% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abalone_data.R
\docType{data}
\name{abalone}
\alias{abalone}
\title{Abalone Data Set}
\format{
A dataframe with 4177 rows and 8 variables
\describe{
\item{Length}{Longest shell measurement in mm}
\item{Diameter}{perpendicular to length in mm}
\item{Height}{with meat in shell in mm}
\item{Whole weight}{whole abalone in grams}
\item{Shucked weight}{weight of meat in grams}
\item{Viscera weight}{gut weight (after bleeding) in grams}
\item{Shell weight}{after being dried in grams}
\item{Rings}{+1.5 gives the age in years}
}
}
\source{
Dua, D. and Graff, C. (2019). UCI Machine Learning Repository \url{http://archive.ics.uci.edu/ml/}. Irvine, CA: University of California, School of Information and Computer Science.
}
\usage{
abalone
}
\description{
The number of rings is the value to predict.
}
\examples{
data(abalone)
}
\keyword{datasets}
