% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{check_is_fitted}
\alias{check_is_fitted}
\title{Fitting Control}
\usage{
check_is_fitted(estimator)
}
\arguments{
\item{estimator}{An estimator with is_fitted attribute}
}
\value{
TRUE if the estimator is fitted, FALSE is not
}
\description{
Checks if the given estimator is fitted
}
