% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial.R
\name{comparison_plot}
\alias{comparison_plot}
\title{Comparsion Plot}
\usage{
comparison_plot(X, y, model_list)
}
\arguments{
\item{X}{Predictors}

\item{y}{Response variables}

\item{model_list}{List of models to be compared}
}
\description{
Plots the fitted functions obtained with various regressors (using their default values) on the
one-dimensional dataset (X, y).
}
