% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{get_functions}
\alias{get_functions}
\title{Get Functions}
\usage{
get_functions()
}
\description{
Prints the available regressors, clustering methods, tree functions and helper functions within LESS package.
}
\examples{
get_functions()
}
