% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{prepareXset}
\alias{prepareXset}
\title{Prepare a Dataset}
\usage{
prepareXset(X)
}
\arguments{
\item{X}{Independent variables}
}
\value{
A named dataframe which consists of X
}
\description{
Takes X dataset and convert it into a dataframe with column names (X_1, X_2 ...)
}
\examples{
X <- matrix(1:20, nrow = 4)
prepareXset(X)
}
