% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressors.R
\name{SklearnEstimator}
\alias{SklearnEstimator}
\title{SklearnEstimator}
\value{
R6 Class of SklearnEstimator
}
\description{
A dummy base R6 class that includes fit, predict functions for estimators
}
\section{Super class}{
\code{\link[less:BaseEstimator]{less::BaseEstimator}} -> \code{SklearnEstimator}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SklearnEstimator-fit}{\code{SklearnEstimator$fit()}}
\item \href{#method-SklearnEstimator-predict}{\code{SklearnEstimator$predict()}}
\item \href{#method-SklearnEstimator-get_type}{\code{SklearnEstimator$get_type()}}
\item \href{#method-SklearnEstimator-get_isFitted}{\code{SklearnEstimator$get_isFitted()}}
\item \href{#method-SklearnEstimator-clone}{\code{SklearnEstimator$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="less" data-topic="BaseEstimator" data-id="get_all_fields"><a href='../../less/html/BaseEstimator.html#method-BaseEstimator-get_all_fields'><code>less::BaseEstimator$get_all_fields()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="BaseEstimator" data-id="get_attributes"><a href='../../less/html/BaseEstimator.html#method-BaseEstimator-get_attributes'><code>less::BaseEstimator$get_attributes()</code></a></li>
<li><span class="pkg-link" data-pkg="less" data-topic="BaseEstimator" data-id="set_random_state"><a href='../../less/html/BaseEstimator.html#method-BaseEstimator-set_random_state'><code>less::BaseEstimator$set_random_state()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SklearnEstimator-fit"></a>}}
\if{latex}{\out{\hypertarget{method-SklearnEstimator-fit}{}}}
\subsection{Method \code{fit()}}{
Dummy fit function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SklearnEstimator$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SklearnEstimator-predict"></a>}}
\if{latex}{\out{\hypertarget{method-SklearnEstimator-predict}{}}}
\subsection{Method \code{predict()}}{
Dummy predict function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SklearnEstimator$predict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SklearnEstimator-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-SklearnEstimator-get_type}{}}}
\subsection{Method \code{get_type()}}{
Auxiliary function returning the type of the class e.g 'estimator'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SklearnEstimator$get_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SklearnEstimator-get_isFitted"></a>}}
\if{latex}{\out{\hypertarget{method-SklearnEstimator-get_isFitted}{}}}
\subsection{Method \code{get_isFitted()}}{
Auxiliary function returning the isFitted flag
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SklearnEstimator$get_isFitted()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SklearnEstimator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SklearnEstimator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SklearnEstimator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
