% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial.R
\name{synthetic_sine_curve}
\alias{synthetic_sine_curve}
\title{Synthetic Sine Curve}
\usage{
synthetic_sine_curve(n_samples = 200)
}
\arguments{
\item{n_samples}{Number of data points to be generated}
}
\description{
A simple function to generate n_samples from sine curve in the range (-10, 10) with some amplitude.
The function returns the dataset (X, y), and plots the function (curve) along with the dataset (circles)
}
