% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial.R
\name{test_timing}
\alias{test_timing}
\title{Compare Fitting Time}
\usage{
test_timing(type = 1, X, y)
}
\arguments{
\item{type}{1 to compare regressors, 2 for comparing classifiers}

\item{X}{Predictors}

\item{y}{Response variables}
}
\description{
Plots a histogram chart which shows the fitting time obtained from various regressors/classifiers (using their default values) on the
given dataset (X, y).
}
\examples{
X <- matrix(sample(100, 20), nrow = 10)
y <- sample(100, 10)
test_timing(1, X, y)
}
